pyproject.toml
[project]
name = "coder-agent"
version = "0.2.0"
description = "ReAct-based coding agent with HumanEval evaluation framework"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.24.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "click>=8.1",
    "pytest>=9.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
coder-agent = "coder_agent.cli.main:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
testpaths = ["tests"]
